AUTONEWART-05 Header module
Fixed menu buttons for IOS. Added aspect ratio. Created callback form component. Added animations.
