+++ package.json
@@ -25,6 +25,7 @@
         "@emotion/react": "^11.9.3",
         "@emotion/styled": "^11.9.3",
         "@mui/material": "^5.8.5",
+        "lottie-web": "^5.9.6",
         "next": "12.1.6",
         "react": "18.2.0",
         "react-dom": "18.2.0",
