+++ src/components/CallbackForm/index.ts
@@ -1,0 +1,3 @@
+import CallbackForm from "./CallbackForm";
+
+export default CallbackForm;
