+++ src/components/CallbackForm/interfaces.ts
@@ -1,0 +1,3 @@
+export interface ICallbackForm {
+  setOpenMenu: (value: boolean) => void;
+}
