+++ src/components/DrawerMenu/DrawerMenu.module.scss
@@ -43,7 +43,7 @@
   flex-direction: column;
   gap: 10px;
   align-items: center;
-  justify-content: center;
+  justify-content: flex-start;
   width: 90%;
   margin: auto;
 }
