+++ src/components/index.ts
@@ -1,3 +1,4 @@
+export { default as CallbackForm } from "./CallbackForm";
 export { default as DrawerMenu } from "./DrawerMenu";
 export { default as HeaderContacts } from "./HeaderContacts";
 export { default as HeaderMenu } from "./HeaderMenu";
